Move Vault code into `external/vault.ts`
