+++ src/external/vault.ts
